+++ ScheduleWidget/Views/NeedsConfigurationView.swift
@@ -1,0 +1,17 @@
+//
+//  NeedsConfigurationView.swift
+//  BsuirScheduleApp
+//
+//  Created by Anton Siliuk on 06/09/2022.
+//  Copyright © 2022 Saute. All rights reserved.
+//
+
+import SwiftUI
+
+struct NeedsConfigurationView: View {
+    var body: some View {
+        Text("widget.needsConfiguration.selectSchedule")
+            .foregroundColor(.secondary)
+            .frame(minWidth: 0, maxWidth: .infinity, minHeight: 0, maxHeight: .infinity)
+    }
+}
