+++ ScheduleWidget/Views/Widgets/ScheduleWidgetEntryMediumView.swift
@@ -1,0 +1,51 @@
+//
+//  ScheduleWidgetEntryMediumView.swift
+//  ScheduleWidgetExtension
+//
+//  Created by Anton Siliuk on 06/09/2022.
+//  Copyright © 2022 Saute. All rights reserved.
+//
+
+import SwiftUI
+import BsuirUI
+import BsuirCore
+
+struct ScheduleWidgetEntryMediumView : View {
+    var entry: Provider.Entry
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 2) {
+            HStack {
+                WidgetDateTitle(date: entry.date)
+                Spacer()
+                ScheduleIdentifierTitle(title: entry.title)
+            }
+
+            switch entry.content {
+            case .needsConfiguration:
+                NeedsConfigurationView()
+            case .pairs(_, []):
+                NoPairsView()
+            case let .pairs(passed, upcoming):
+                let pairs = PairsToDisplay(
+                    passed: passed,
+                    upcoming: upcoming,
+                    maxVisibleCount: 2
+                )
+
+                VStack(alignment: .leading, spacing: 4) {
+                    ForEach(pairs.visible) { pair in
+                        PairView<EmptyView>(pair: pair, isCompact: true)
+                    }
+                }
+                .padding(.top, 6)
+
+                Spacer(minLength: 0)
+
+                RemainingPairsView(pairs: pairs.upcomingInvisible, visibleCount: 3, showTime: .first)
+            }
+        }
+        .padding()
+        .background(Color(.systemBackground))
+    }
+}
